move file count limit from dvObject to dvObjectContainer
