+++ src/main/resources/db/migration/V6.5.0.14.sql
@@ -1,1 +1,2 @@
-ALTER TABLE DVOBJECT ADD COLUMN IF NOT EXISTS datasetfilecountlimit bigint;
+ALTER TABLE DATASET ADD COLUMN IF NOT EXISTS datasetfilecountlimit bigint;
+ALTER TABLE DATAVERSE ADD COLUMN IF NOT EXISTS datasetfilecountlimit bigint;
